requirements.txt
pipecat-ai[assemblyai,silero,daily,webrtc]
python-dotenv
fastapi
uvicorn
httpx
aiohttp
